+++ src/engine_v0_2/src/RectangularNonPlayer.cpp
@@ -9,11 +9,11 @@
 
 using namespace ge;
 
-RectangularNonPlayer::RectangularNonPlayer( float width, float height, eColor baseColor, int fuzz,Movement * movement, float lifetime, const sf::Vector2f& initPosition ) :
+RectangularNonPlayer::RectangularNonPlayer( float width, float height, eColor baseColor, int fuzz,Movement * movement, unsigned long long lifetime, const sf::Vector2f& initPosition ) :
                                             NonPlayer( new sf::RectangleShape( sf::Vector2f( width, height ) ) , baseColor, fuzz, movement, lifetime, initPosition ), m_width( width ), m_height( height ) {
 }
 
-RectangularNonPlayer::RectangularNonPlayer( float width, float height, const ge::Color& color,Movement * movement, float lifetime, const sf::Vector2f& initPosition ) : 
+RectangularNonPlayer::RectangularNonPlayer( float width, float height, const ge::Color& color,Movement * movement, unsigned long long lifetime, const sf::Vector2f& initPosition ) : 
                                             NonPlayer( new sf::RectangleShape( sf::Vector2f( width, height ) ) , color, movement, lifetime, initPosition ), m_width( width ), m_height( height ){
     
 }
